Fix a bug in checks for table entries
RuntimeAPI contains some instances where it is checked if an entry
exists by branching on the truth value the entry handle variable.
In the case of an entry_handle that is equal to 0, these led to an
exception raised, even though the entry exists in the table.

This is fixed by checking explicitly if the the handle is None.